Refactor the entire. Make into one
