+++ main.go
@@ -2,8 +2,8 @@ package main
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
-	"io"
 	"log"
 	"os"
 	"path/filepath"
@@ -12,20 +12,15 @@ import (
 	clilog "github.com/b4b4r07/go-cli-log"
 	"github.com/jessevdk/go-flags"
 	"gopkg.in/src-d/go-git.v4"
+	"gopkg.in/src-d/go-git.v4/plumbing"
 	"gopkg.in/src-d/go-git.v4/plumbing/object"
+	"gopkg.in/src-d/go-git.v4/utils/merkletrie"
 )
 
 var (
 	Version  = "unset"
 	Revision = "unset"
 )
-
-type CLI struct {
-	Option Option
-	Stdout io.Writer
-	Stderr io.Writer
-	Repo   *git.Repository
-}
 
 type Option struct {
 	Filters       []string `long:"filter" description:"Filter the kind of changed objects" default:"all" choice:"added" choice:"modified" choice:"deleted" choice:"all"`
@@ -40,10 +35,13 @@ type Option struct {
 }
 
 func main() {
-	os.Exit(run(os.Args[1:]))
+	if err := run(os.Args[1:]); err != nil {
+		fmt.Fprintf(os.Stderr, "[ERROR] %v\n", err)
+		os.Exit(1)
+	}
 }
 
-func run(args []string) int {
+func run(args []string) error {
 	clilog.Env = "LOG"
 	clilog.SetOutput()
 	defer log.Printf("[INFO] finish main function")
@@ -54,25 +52,12 @@ func run(args []string) int {
 	var opt Option
 	args, err := flags.ParseArgs(&opt, args)
 	if err != nil {
-		return 1
+		return err
 	}
-	cli := CLI{
-		Option: opt,
-		Stdout: os.Stdout,
-		Stderr: os.Stderr,
-		Repo:   nil,
-	}
-	if err := cli.Run(args); err != nil {
-		fmt.Fprintf(os.Stderr, "[ERROR] %v\n", err)
-		return 1
-	}
-	return 0
-}
 
-func (c *CLI) Run(args []string) error {
 	switch {
-	case c.Option.Version:
-		fmt.Fprintf(c.Stdout, "%s (%s)\n", Version, Revision)
+	case opt.Version:
+		fmt.Fprintf(os.Stdout, "%s (%s)\n", Version, Revision)
 		return nil
 	}
 
@@ -85,16 +70,14 @@ func (c *CLI) Run(args []string) error {
 	if err != nil {
 		return err
 	}
-	repo := abs
-	log.Printf("[INFO] git repo: %s", repo)
+	log.Printf("[INFO] git repo: %s", abs)
 
-	r, err := git.PlainOpen(repo)
+	repo, err := git.PlainOpen(abs)
 	if err != nil {
 		return fmt.Errorf("cannot open repository: %w", err)
 	}
-	c.Repo = r
 
-	branch, err := c.GetCurrentBranchFromRepository()
+	branch, err := GetCurrentBranchFromRepository(repo)
 	if err != nil {
 		return err
 	}
@@ -103,49 +86,49 @@ func (c *CLI) Run(args []string) error {
 	var base *object.Commit
 
 	switch branch {
-	case c.Option.DefaultBranch:
+	case opt.DefaultBranch:
 		log.Printf("[DEBUG] Getting previous HEAD commit")
-		prev, err := c.previousCommit()
+		prev, err := previousCommit(repo)
 		if err != nil {
 			return err
 		}
 		base = prev
 	default:
 		log.Printf("[DEBUG] Getting remote commit")
-		remote, err := c.remoteCommit("origin/" + c.Option.DefaultBranch)
+		remote, err := remoteCommit("origin/"+opt.DefaultBranch, repo)
 		if err != nil {
 			return err
 		}
 		base = remote
 	}
 
-	if c.Option.MergeBase != "" {
+	if opt.MergeBase != "" {
 		log.Printf("[DEBUG] Comparing with merge-base")
-		h, err := c.Repo.Head()
+		h, err := repo.Head()
 		if err != nil {
 			return err
 		}
 		currentBranch := h.Name().Short()
-		base, err = c.mergeBase(c.Option.MergeBase, currentBranch)
+		base, err = mergeBase(opt.MergeBase, currentBranch, repo)
 		if err != nil {
 			return err
 		}
 	}
 
 	log.Printf("[DEBUG] Getting current commit")
-	current, err := c.currentCommit()
+	current, err := currentCommit(repo)
 	if err != nil {
 		return err
 	}
 
-	stats, err := c.getStats(base, current)
+	stats, err := getStats(base, current, repo)
 	if err != nil {
 		return err
 	}
 
-	log.Printf("[INFO] Option filters: %#v", c.Option.Filters)
+	log.Printf("[INFO] Option filters: %#v", opt.Filters)
 	var ss Stats
-	for _, filter := range c.Option.Filters {
+	for _, filter := range opt.Filters {
 		switch filter {
 		case "all":
 			ss = stats
@@ -180,20 +163,20 @@ func (c *CLI) Run(args []string) error {
 		stats = ss
 	}
 
-	if c.Option.Dirname {
+	if opt.Dirname {
 		stats = stats.Map(func(stat Stat) Stat {
 			stat.Path = stat.Dir
 			return stat
 		})
 	}
 
-	if c.Option.DirExist {
+	if opt.DirExist {
 		stats = stats.Filter(func(stat Stat) bool {
 			return stat.DirExist
 		})
 	}
 
-	if c.Option.DirNotExist {
+	if opt.DirNotExist {
 		stats = stats.Filter(func(stat Stat) bool {
 			return !stat.DirExist
 		})
@@ -201,28 +184,28 @@ func (c *CLI) Run(args []string) error {
 
 	stats = stats.Unique()
 
-	switch c.Option.Output {
+	switch opt.Output {
 	case "json":
 		r := struct {
 			Repo  string `json:"repo"`
 			Stats Stats  `json:"stats"`
 		}{
-			Repo:  repo,
+			Repo:  abs,
 			Stats: stats,
 		}
-		return json.NewEncoder(c.Stdout).Encode(&r)
+		return json.NewEncoder(os.Stdout).Encode(&r)
 	case "":
 		for _, stat := range stats {
-			fmt.Fprintln(c.Stdout, stat.Path)
+			fmt.Fprintln(os.Stdout, stat.Path)
 		}
 	default:
-		return fmt.Errorf("%s: invalid output format", c.Option.Output)
+		return fmt.Errorf("%s: invalid output format", opt.Output)
 	}
 
 	return nil
 }
 
-func (c CLI) getStats(from, to *object.Commit) (Stats, error) {
+func getStats(from, to *object.Commit, repo *git.Repository) (Stats, error) {
 	src, err := to.Tree()
 	if err != nil {
 		return Stats{}, err
@@ -242,7 +225,7 @@ func (c CLI) getStats(from, to *object.Commit) (Stats, error) {
 
 	var stats []Stat
 	for _, change := range changes {
-		stat, err := c.fileStatsFromChange(change)
+		stat, err := fileStatsFromChange(change, repo)
 		if err != nil {
 			continue
 		}
@@ -251,4 +234,239 @@ func (c CLI) getStats(from, to *object.Commit) (Stats, error) {
 	}
 
 	return stats, nil
+}
+
+func currentCommit(repo *git.Repository) (*object.Commit, error) {
+	ref, err := repo.Head()
+	if err != nil {
+		return nil, err
+	}
+
+	log.Printf("[DEBUG] %s: get commit", ref.Name().String())
+	return repo.CommitObject(ref.Hash())
+}
+
+func previousCommit(repo *git.Repository) (*object.Commit, error) {
+	hash, err := repo.ResolveRevision("HEAD^")
+	if err != nil {
+		return nil, err
+	}
+
+	return repo.CommitObject(*hash)
+}
+
+func remoteCommit(name string, repo *git.Repository) (*object.Commit, error) {
+	refs, err := repo.References()
+	if err != nil {
+		return nil, err
+	}
+
+	var cmt *object.Commit
+	err = refs.ForEach(func(ref *plumbing.Reference) error {
+		if ref.Name().String() == fmt.Sprintf("refs/remotes/%s", name) {
+			commit, err := repo.CommitObject(ref.Hash())
+			if err != nil {
+				return err
+			}
+			log.Printf("[DEBUG] %s: get commit", ref.Name().String())
+			cmt = commit
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return cmt, nil
+}
+
+func masterCommit(name string, repo *git.Repository) (*object.Commit, error) {
+	branches, err := repo.Branches()
+	if err != nil {
+		return nil, err
+	}
+
+	var cmt *object.Commit
+	err = branches.ForEach(func(branch *plumbing.Reference) error {
+		if branch.Name().String() == fmt.Sprintf("refs/heads/%s", name) {
+			commit, err := repo.CommitObject(branch.Hash())
+			if err != nil {
+				return err
+			}
+			log.Printf("[DEBUG] %s: get commit", branch.Name().String())
+			cmt = commit
+		}
+		return nil
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	return cmt, nil
+}
+
+func fileStatsFromChange(change *object.Change, repo *git.Repository) (Stat, error) {
+	action, err := change.Action()
+	if err != nil {
+		return Stat{}, err
+	}
+
+	var kind Kind
+	var path string
+	switch action {
+	case merkletrie.Delete:
+		kind = Deletion
+		path = change.From.Name
+	case merkletrie.Insert:
+		kind = Addition
+		path = change.To.Name
+	case merkletrie.Modify:
+		kind = Modification
+		path = change.To.Name
+	default:
+		kind = Unknown
+	}
+
+	return Stat{
+		Kind: kind,
+		Path: path,
+		File: path,
+		Dir:  filepath.Dir(path),
+		DirExist: func() bool {
+			_, err := os.Stat(filepath.Dir(path))
+			return err == nil
+		}(),
+	}, nil
+}
+
+// https://github.com/go-git/go-git/blob/master/_examples/merge_base/main.go
+func mergeBase(baseRev, commitRev string, repo *git.Repository) (*object.Commit, error) {
+	log.Printf("[DEBUG] baseRev: %s, commitRev: %s", baseRev, commitRev)
+
+	// Get the hashes of the passed revisions
+	var hashes []*plumbing.Hash
+	for _, rev := range []string{baseRev, commitRev} {
+		hash, err := repo.ResolveRevision(plumbing.Revision(rev))
+		if err != nil {
+			return nil, err
+		}
+		hashes = append(hashes, hash)
+	}
+
+	// Get the commits identified by the passed hashes
+	var commits []*object.Commit
+	for _, hash := range hashes {
+		commit, err := repo.CommitObject(*hash)
+		if err != nil {
+			return nil, err
+		}
+		commits = append(commits, commit)
+	}
+
+	res, err := commits[0].MergeBase(commits[1])
+	if err != nil {
+		return nil, err
+	}
+
+	if len(res) == 0 {
+		return nil, errors.New("failed to get merge-base")
+	}
+
+	return res[0], nil
+}
+
+// https://github.com/src-d/go-git/issues/1030
+func GetCurrentBranchFromRepository(repo *git.Repository) (string, error) {
+	branchRefs, err := repo.Branches()
+	if err != nil {
+		return "", err
+	}
+
+	headRef, err := repo.Head()
+	if err != nil {
+		return "", err
+	}
+
+	var currentBranchName string
+	err = branchRefs.ForEach(func(branchRef *plumbing.Reference) error {
+		if branchRef.Hash() == headRef.Hash() {
+			currentBranchName = branchRef.Name().Short()
+			return nil
+		}
+
+		return nil
+	})
+	if err != nil {
+		return "", err
+	}
+
+	return currentBranchName, nil
+}
+
+type Kind int
+
+const (
+	Addition Kind = iota
+	Deletion
+	Modification
+	Unknown
+)
+
+func (k Kind) String() string {
+	switch k {
+	case Addition:
+		return "insert"
+	case Deletion:
+		return "delete"
+	case Modification:
+		return "modify"
+	default:
+		return "unknown"
+	}
+}
+
+func (k Kind) MarshalJSON() ([]byte, error) {
+	return json.Marshal(k.String())
+}
+
+// Stat represents the stats for a file in a commit.
+type Stat struct {
+	Kind     Kind   `json:"kind"`
+	Path     string `json:"path"`
+	File     string `json:"file"`
+	Dir      string `json:"dir"`
+	DirExist bool   `json:"dir-exist"`
+	Output   string `json:"-"`
+}
+
+type Stats []Stat
+
+func (ss *Stats) Filter(f func(Stat) bool) Stats {
+	stats := make([]Stat, 0)
+	for _, stat := range *ss {
+		if f(stat) {
+			stats = append(stats, stat)
+		}
+	}
+	return stats
+}
+
+func (ss *Stats) Map(f func(Stat) Stat) Stats {
+	stats := make([]Stat, len(*ss))
+	for i, stat := range *ss {
+		stats[i] = f(stat)
+	}
+	return stats
+}
+
+func (ss *Stats) Unique() Stats {
+	m := make(map[string]bool)
+	stats := make([]Stat, 0)
+	for _, stat := range *ss {
+		if !m[stat.Path] {
+			m[stat.Path] = true
+			stats = append(stats, stat)
+		}
+	}
+	return stats
 }
